finish refactoring io code
